+++ routes/api/contacts.js
@@ -1,145 +1,12 @@
 const express = require("express");
 const router = express.Router();
-const { nanoid } = require("nanoid");
-const Joi = require("joi");
 
-const contactsActions = require("../../utils/utils");
-const HttpError = require("../../helpers/HttpError");
+const ctrl = require("../../controllers/contacts");
 
-const PostSchema = Joi.object({
-  name: Joi.string().min(6).max(30).required(),
-  email: Joi.string().email().required(),
-  phone: Joi.string()
-    .pattern(/^[+]?[(]?[0-9]{3}[)]?[-\s.]?[0-9]{3}[-\s.]?[0-9]{4,6}$/)
-    .required(),
-});
-
-const PutSchema = Joi.object({
-  name: Joi.string().min(6).max(30),
-  email: Joi.string().email(),
-  phone: Joi.string().pattern(
-    /^[+]?[(]?[0-9]{3}[)]?[-\s.]?[0-9]{3}[-\s.]?[0-9]{4,6}$/
-  ),
-});
-
-router.get("/", async (req, res, next) => {
-  try {
-    const getAll = await contactsActions.listContacts();
-    res.json({
-      status: "success",
-      code: 200,
-      data: {
-        getAll,
-      },
-    });
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.get("/:contactId", async (req, res, next) => {
-  try {
-    const id = req.params.contactId;
-    const getById = await contactsActions.getContactById(id);
-    if (JSON.stringify(getById) === "[]") {
-      throw HttpError(404, "Not found");
-    } else {
-      res.json({
-        status: "success",
-        code: 200,
-        data: {
-          getById,
-        },
-      });
-    }
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.post("/", async (req, res, next) => {
-  try {
-    const { name, email, phone } = req.body;
-
-    const { error } = PostSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-
-    if (!name || !email || !phone) {
-      throw HttpError(400, "missing required name field");
-    } else {
-      const contact = {
-        id: nanoid(),
-        name: name,
-        email: email,
-        phone: phone,
-      };
-
-      const add = await contactsActions.addContact(contact);
-
-      res.json({
-        status: "created",
-        code: 201,
-        data: {
-          add,
-        },
-      });
-    }
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.delete("/:contactId", async (req, res, next) => {
-  try {
-    const id = req.params.contactId;
-    const remove = await contactsActions.removeContact(id);
-
-    if (JSON.stringify(remove) === "[]") {
-      throw HttpError(404, "Not found");
-    } else {
-      res.json({
-        status: "contact deleted",
-        code: 200,
-        data: {
-          remove,
-        },
-      });
-    }
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.put("/:contactId", async (req, res, next) => {
-  try {
-    const id = req.params.contactId;
-
-    const { error } = PutSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-
-    if (Object.keys(req.body).length === 0) {
-      throw HttpError(400, "missing required name field");
-    } else {
-      const update = await contactsActions.updateContact(id, req.body);
-      if (JSON.stringify(update) === "[]") {
-        throw HttpError(404, "Not found");
-      } else {
-        res.json({
-          status: "success",
-          code: 200,
-          data: {
-            update,
-          },
-        });
-      }
-    }
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/", ctrl.getAll);
+router.get("/:contactId", ctrl.getById);
+router.post("/", ctrl.add);
+router.delete("/:contactId", ctrl.remove);
+router.put("/:contactId", ctrl.update);
 
 module.exports = router;
